+++ app/src/main/java/com/fyd/miku/model/pmd/FaceMorph.java
@@ -10,7 +10,7 @@ public class FaceMorph {
     String morphName;       //动画名称
     String morphNameEnglish;
     int verticesCount;      //动画影响的顶点数量
-    byte morphType;          //动画类 0.base 1. 眉毛动画，2. 眼睛动画， 3嘴唇动画 4.其他
+    int morphType;          //动画类 0.base 1. 眉毛动画，2. 眼睛动画， 3嘴唇动画 4.其他
     List<Vertex> vertices;
 
     FaceMorph() {
